sq-poller: use the issued command timestamp to workout bootup timestamp
We previously worked out the boot timestamp as:
bootup_timestamp = received_response_ts - uptime

But the handling of a received response could be delayed by the
processing of some data and this might cause some false reboot
detection. To solve this issue we use instead the timestamp when the
commands has been issued, in this way the time drift is given by the
time that the node requires to receive anb execute the command. Which
should be less variable.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>